% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download.R
\name{mod_download_ui}
\alias{mod_download_ui}
\alias{mod_download_server_core}
\title{download UI Function}
\usage{
mod_download_ui(id)

mod_download_server_core(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\keyword{internal}
