% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter_ui}
\alias{mod_filter_ui}
\alias{mod_filter_server}
\title{filter UI Function}
\usage{
mod_filter_ui(id)

mod_filter_server(id, dataSet = "", probes = "")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\keyword{internal}
