% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap.R
\name{mod_heatmap_ui}
\alias{mod_heatmap_ui}
\alias{mod_heatmap_server}
\title{heatmap UI Function}
\usage{
mod_heatmap_ui(id)

mod_heatmap_server(id, normal = "", group_list = "", filter = "")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\keyword{internal}
