% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal.R
\name{mod_normal_ui}
\alias{mod_normal_ui}
\alias{mod_normal_server}
\title{normal UI Function}
\usage{
mod_normal_ui(id)

mod_normal_server(id, group_list = "", filter = "")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\keyword{internal}
