% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pdata.R
\name{mod_pdata_ui}
\alias{mod_pdata_ui}
\alias{mod_pdata_server}
\title{pdata UI Function}
\usage{
mod_pdata_ui(id)

mod_pdata_server(id, dataSet = "")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\keyword{internal}
