% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_probe.R
\name{mod_probe_ui}
\alias{mod_probe_ui}
\alias{mod_probe_server}
\title{probe UI Function}
\usage{
mod_probe_ui(id)

mod_probe_server(id, dataSet = "")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\keyword{internal}
